+++ frontend/src/components/CardInstancia.vue
@@ -20,10 +20,7 @@ defineProps({
 });
 </script>
 <template>
-	<a class="flex flex-col justify-between items-center w-[240px] h-[240px] p-4 border border-zinc-700
-        rounded-md shadow-md transition duration-300 cursor-pointer bg-zinc-950 
-        hover:bg-gradient-to-br from-zinc-950 to-zinc-800" :href="`${url}`"
-	>
+	<a class="flex flex-col justify-between items-center w-[240px] h-[240px] p-4 border border-zinc-700 rounded-md shadow-md transition duration-300 cursor-pointer bg-zinc-950 hover:bg-gradient-to-br from-zinc-950 to-zinc-800" :href="`${url}`">
 		<header class="flex flex-col items-start text-start w-full">
 			<h1 class="text-lg font-semibold">{{ titulo }}</h1>
 			<h2 class="text-sm font-medium text-zinc-300">{{ semestre }}</h2>
