+++ frontend/src/components/Login.vue
@@ -1,7 +1,8 @@
 <template>
     <section class="flex justify-center items-center w-full h-full">
-        <button
-            class="bg-zinc-50 text-zinc-900 font-medium py-2 px-4 rounded-md hover:bg-zinc-300 transition duration-300 cursor-pointer shadow-md"
-            onclick="window.location.href = '/instancias'">Ingresar</button>
+        <!-- Botón de acceso -->
+        <button class="bg-zinc-50 text-zinc-900 font-medium py-2 px-4 rounded-md hover:bg-zinc-300 transition duration-300 cursor-pointer shadow-md" onclick="window.location.href = '/instancias'">
+            Ingresar
+        </button>
     </section>
 </template>
