+++ frontend/src/components/Navbar.vue
@@ -1,8 +1,10 @@
 <template>
     <nav class="w-full flex justify-between items-center p-4 bg-zinc-900 border-b border-zinc-800 shadow-md">
+        <!-- Lado izquierdo -->
         <div class="flex items-center">
             <a href="/" class="font-bold text-lg">Ruleta Aleatoria</a>
         </div>
+        <!-- Lado derecho -->
         <div class="flex justify-center items-center gap-4">
             <a href="/" class="font-medium text-md">About</a>
             <a href="/" class="font-medium text-md">Contact</a>
